refactor: move get all published posts query function to main module
