+++ src/modules/posts/post.services.ts
@@ -37,26 +37,6 @@ export const getAllPosts = async (userId: string, page: number) => {
   return posts
 }
 
-export const getAllPublishedPosts = async (page: number) => {
-  const PER_PAGE_ITEMS = 5
-
-  const posts = await db.post.findMany({
-    where: { postStatus: 'PUBLISHED' },
-    include: {
-      author: {
-        select: {
-          name: true,
-        },
-      },
-    },
-    take: PER_PAGE_ITEMS,
-    skip: (page - 1) * PER_PAGE_ITEMS,
-    orderBy: { createdAt: 'desc' },
-  })
-
-  return posts
-}
-
 export const getAllPostsCount = async (userId: string) => {
   const totalPosts = await db.post.count({
     where: { authorId: userId },
